Add check for subtracted sets (#59)
Currently, the workflow does not run a check for a case where a pull request removes an entire set from the list. A check should therefore be added ensuring that the well-known page for the primary of a removed set serves an error 404. This will confirm that the set was intended to be removed from the list by the owner of the affected domain.